r101: disabling coverage-based depth by default
In the old version, the depth cut-off of 2nd round read cleaning depends on the
estimated coverage in the first round. There are two potential issues. Firstly,
the coverage estimate is very crude. This may affect the accuracy of the cutoff.
Secondly and more importantly, when we have mixed samples with different
averaged coverage (e.g. metagenomics or pooling), the coverage varies with
biospecimen. It is not right to use a single coverage cutoff.

This revision switches off the coverage-based cutoff. We may completely disable
this "feature" in future.